refactor: change project structure to go-standard
